requirements.txt
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.107.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.20.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
flask>=3.0.0
python-docx>=1.1.0
GoogleNews>=1.6.14